requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
streamlit>=1.33.0

# ML
xgboost>=2.0.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0

# LLMs + Embeddings
sentence-transformers>=2.5.1
transformers>=4.40.0
torch>=2.2.0

# Vectorstore
faiss-cpu>=1.7.4

# Explainable AI
shap>=0.45.0

# Utility
tqdm>=4.66.0
python-dotenv>=1.0.1
